Added redirect away from auth pages if user is logged in
Added navbar detection of whether or not the user is logged in, if so displays correct navigation menu.

Added logout button to my account nav for logged in users

checking pathname changes on useEffect and toggling isMyAccountNavOpen to false on pathname change is amazing, no need for a function to set isMyAccountNavOpen to false on every link click, etc.